@@ -1,15 +1,14 @@
-import {Component, OnInit} from '@angular/core';
-import {AuthService} from './shared/services/auth.service';
+import { Component, OnInit } from '@angular/core';
+import { AuthService } from './shared/services';
 
 @Component({
   selector: 'app-root',
   template: '<router-outlet></router-outlet>'
 })
 export class AppComponent implements OnInit {
-  constructor(private auth: AuthService) {
-  }
+  constructor(private auth: AuthService) { }
 
-  ngOnInit() {
+  ngOnInit(): void {
     const potentialToken = localStorage.getItem('auth-token');
     if (potentialToken !== null) {
       this.auth.setToken(potentialToken);
