@@ -1,30 +1,23 @@
-import { ClientsViewComponent } from './shared/layouts/clients-layout/clients-view/clients-view.component';
-import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
+import { ReactiveFormsModule } from '@angular/forms';
+import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+// import { StoreModule } from '@ngrx/store';
 
-
-import { AppStoreModule } from './store/app-store.module';
-
-import { TimepickerModule } from 'ngx-bootstrap/timepicker';
-import { BsDatepickerModule } from 'ngx-bootstrap/datepicker';
-
-import { AppComponent } from './app.component';
-import { LoginPageComponent } from './login-page/login-page.component';
-import { AuthLayoutComponent } from './shared/layouts/auth-layout/auth-layout.component';
 import { AppRoutingModule } from './app-routing.module';
-import { SiteLayoutComponent } from './shared/layouts/site-layout/site-layout.component';
-import { RegisterPageComponent } from './register-page/register-page.component';
-import { TokenInterceptor } from './shared/classes/token.interceptor';
-
+import { AppComponent } from './app.component';
 import { ClientsPageComponent } from './clients-page/clients-page.component';
 import { TasksComponent } from './clients-page/tasks/tasks.component';
+import { LoginPageComponent } from './login-page/login-page.component';
 import { MonthlyReportsComponent } from './monthly-reports/monthly-reports.component';
-import { ClientStatisticComponent } from './shared/layouts/clients-layout/clients-statistic.component';
-import { SharedModule } from './shared/module/shared-module/shared-module.module';
-
+import { RegisterPageComponent } from './register-page/register-page.component';
+import { AuthLayoutComponent } from './shared/layouts/auth-layout/auth-layout.component';
+import {
+  ClientStatisticComponent
+} from './shared/layouts/clients-layout/clients-statistic.component';
+import { SiteLayoutComponent } from './shared/layouts/site-layout/site-layout.component';
+import { SharedModule } from './shared/module';
+// import { AppStoreModule } from './store/app-store.module';
 
 @NgModule({
   declarations: [
@@ -37,27 +30,17 @@ import { SharedModule } from './shared/module/shared-module/shared-module.module
     TasksComponent,
     MonthlyReportsComponent,
     ClientStatisticComponent,
-    ClientsViewComponent,
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    FormsModule,
-    SharedModule,
-    AppStoreModule,
-    ReactiveFormsModule,
-    HttpClientModule,
+    // AppStoreModule,
     BrowserAnimationsModule,
-    TimepickerModule.forRoot(),
-    BsDatepickerModule.forRoot(),
+    ReactiveFormsModule,
+    SharedModule,
+    // StoreModule.forRoot({}, {})
   ],
-  providers: [
-    {
-      provide: HTTP_INTERCEPTORS,
-      multi: true,
-      useClass: TokenInterceptor
-    }
-  ],
+  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
