@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { Store } from '@ngrx/store';
 
 import { GettingAllTasks } from '../store/actions/tasks.action';
-import { AppState } from '../store/app-store.module';
+// import { AppState } from '../store/app-store.module';
 import { TasksService } from '../shared/services/tasks.service';
 
 @Component({
@@ -17,9 +17,11 @@ import { TasksService } from '../shared/services/tasks.service';
   ]
 })
 export class ClientsPageComponent {
-  constructor(private taskService: TasksService, private store: Store<AppState>, ) {
+  constructor(private taskService: TasksService, 
+    // private store: Store<AppState>
+     ) {
     this.taskService.getAllClientTask().subscribe(res => {
-      this.store.dispatch(new GettingAllTasks(res));
-    })
+      // this.store.dispatch(new GettingAllTasks(res));
+    });
   }
 }
